tests: Modernize siteselector QUnit tests with async-await
== What ==

Remove local complexity in favor of native QUnit and JavaScript
idioms so that there is no need to trust or debug the test suite
itself.

Keep only one control flow for assertions: the QUnit.test function
itself. Use source-controlled callbacks like .one() only to collect
data, not to decide whether or when to perform an assertion because
that means if it is not called, the test passes anyway.

The async-done workaround solves this mostly, but by failing through
a timeout rather than something quicker and more explicit with a
partial result that can inform you where/what part got stuck

== Why ==

At some point, this test failed for me locally and it was stuck
forever instead of showing an error or timing out. The reason was
that done() was never called in some cases, so the test just gets
stuck with no information about why or where it is stuck.

See also https://qunitjs.com/api/assert/verifySteps/ for more
behind this idea.

Change-Id: Ia638daba805fc7e31e7a105aaafc2f26e13e2e53