Fix QUnit test when cldr extension is not enabled
wikibase.Wikibase.getMonolingualTextLanguages() test fails when cldr is
not present: it is expecting a language `abe` to be present which is
defined by cldr.

Replace cldr specific language `abe` by `aeb` which is defined in
MediaWiki core.
Update the assertion to show the language code being asserted.

This let us run the repository QUnit tests without `cldr` installed.

Bug: T389998
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: If83f8d74fff8b9a97ad2e30d080d7eae1fc553cb