Use shared mw.Api() instance
Later, we can add shared parameters that all requests should use here.
It’s also somewhat useful for the tests – it simplifies the mocking in
propertySelector.spec.js, as we no longer have to track the mockedGet.

In preparation for change I6760a0ea8d, already mock the API in several
other tests which end up importing it; this isn’t yet necessary in this
change, but becomes necessary in that one, as otherwise the creation of
the API fails. We should probably move these mocks – Codex, icons, API,
and perhaps supportedDatatypes too – to a shared place soon.

Bug: T401405
Change-Id: I1a7cc609c1b5a86cefbe7d7613ab0c88ece3b22f