Use wbparsevalue when editing statement
Instead of directly using the text input value as the data value, send
the user input to the API and use the returned data value. One visible
benefit of this change is that user input with leading or trailing
whitespace will now be silently trimmed by wbparsevalue, rather than
result in an error when we send it untrimmed to wbeditentity.

This is kind of the most bare-bones version of this change, only wiring
the feature into the editStatementsStore. Future improvements for
follow-up changes include, but are not limited to:

- kick off the wbparsevalue requests as soon as the user types ("input"
  event, or perhaps "change" event?), instead of only on save
- disable the submit button while the wbparsevalue request is pending
  (the "resolved" property is meant to make that available synchronously)
- disable the submit button if the parsed value is the same as the
  current value (no change, e.g. only whitespace changes)
- disable the submit button if parsing the value failed (or show an
  error message in that case? try again after a backoff?)

In the editStatementsStore, I moved statementsForSerialization and
buildStatementObjectFromMutableStatement into the actions so that they
can access the right store even though there are now potentially
`await`s between the different store calls.

Bug: T405238
Change-Id: I66077e99305d5da3284713b838dadd74869164ea