Try setting the value a second time in browser test
So far I haven’t been able to figure out why the value is sometimes not
set, and nobody else seems to have had any ideas either. Let’s just
(ab)use the browser.keys() command as a “wait for the page to have fully
initialized itself”, and then try to set the value again, if necessary.
(To ensure that the original bug is still tested, also re-type the
summary then, so that the stashedit API request can be sent again.)

Bug: T388228
Change-Id: Icb48eaf23d6a0e031fdd66134da9f61c5f92cb3d