GQL: Add initial GraphQL server
This exposes the GraphQL server on Special:WikibaseGraphQL. All it can
do for now is repeating an Item ID back to the user. Actual data
fetching will be added in a follow-up patch.

We're not testing the special page itself here yet because it's awkward
while webonyx/graphql-php is still awaiting security review and
unavailable in the vendor CI jobs. We have an integration test for
GraphQLService and will need to check the special page manually for now.

Bug: T404832
Change-Id: I6640be702eb3c674631eda4ab0e0dbdc7b153a43