Align qualifier snaks
Makes the spacing of qualifier properties and values consistent in
a two-column format. The property label is a fixed width. When the
text is longer than this width, it is truncated with an ellipsis.

Note: qualifiers don't look quite right when editing a statement,
but this should be considered out of scope. In T405739, existing
qualifiers will become editable, changing the look significantly.

Bug: T404991
Change-Id: Ia6a1206a8d7724be30cf9bc3b8376a3258d7a369