Use assert/assertuser in MEX
When making an edit, make sure the user is still logged in using the
assertCurrentUser() method [1], which adds the assert and/or assertuser
parameters to make the API request fail if the user state is no longer
the same as when the page was loaded.

Note that the possible resulting error isn’t yet handled at all. A basic
error message is added in change I987ea3c052; further error handling
will happen later.

[1]: https://doc.wikimedia.org/mediawiki-core/master/js/mw.Api.html

Change-Id: I62313b38ebb0dc3dfca9d19df1cadd95d10fda4a