Parse values after 300 ms without waiting for publish
Just like the desktop UI, parse values as soon as the user types them,
debounced with a delay of 300 ms. Right now this just speeds up the
saving/publishing a little bit (by the time the user clicks the button,
we may already have finished parsing the value, so we can send the
actual edit immediately), but it’s also a building block for further
improvements in T405238 – such as change I6706307300.

Only applies to the main snak / value for now. Qualifiers and references
aren’t using wbparsevalue yet, as far as I can tell (but they should).

Bug: T405238
Change-Id: Iaecf42a89e53900ee010deacce586bd678a48520