Disable edit button until statement value is parsed
New getters isFullyParsed in the editStatement(s)Store check if all the
data of the statements is available in the parsedValueStore, and we
disable the button based on that.

For this to work, the parsedValueStore’s parsedValue needs to be made
reactive() – ideally that would’ve happened in change I66077e9930
(commit 7841f575a5) but I didn’t realize it at the time.

Bug: T405238
Change-Id: I67063073007ddf0d90e5f64199f5f6858e78457f