Populate parsedValueStore with initial statements
This is important for change I6706307300 – if we disable the “publish”
button when values haven’t been parsed yet, but we don’t have a parsed
value yet (and aren’t going to send a parse request for the initial
input either), then the “publish” button stays disabled until users edit
the value, even though they might not want to edit the value (e.g. they
might only want to change the rank). We could also solve this by sending
a parse request for the initial value, but IMHO that’s wasteful.

This assumes that, for any property with data type string, the value of
a saved data value would parse to that same data value again; this
implies that, whenever parsing the value results in a data value without
a string, Wikibase would not have allowed saving the original string. As
far as I know, this is true – string value parsing only does minor
transformations such as trimming whitespace and perhaps normalizing
Unicode (not sure), and untrimmed values cannot be saved.

For now, apply this only to snaks with the data type string. It would
probably be possible to also apply this to most, perhaps all, other data
types with the data *value* type string (i.e., data types URL, external
identifier, tabular data, musical notation, …), but as we don’t support
editing these yet, I don’t think it’s necessary for now. And for data
types with non-string data value types (entity IDs, monolingual text,
quantities, globe coordinates, time), I’m not yet sure if we’ll use
wbparsevalue at all.

The peekParsedValue action (really a “getter with parameters”, but I
think it’s cleaner to write it as an action than a getter returning a
function) is added here for the tests, but is also subsequently used in
change I6706307300.

Bug: T405238
Change-Id: I6fcd7c2359b0f358b864518e19b57446e5092143