Edit and delete snaks on references
Makes snaks for existing and new references editable when editing a
statement. Adding new snaks to references is not covered in this
commit, and will instead by done in a followup (If35fe93).

Error handling and input validation will be added + improved in a
followup patch or future ticket, as this affects all inputs.

Included in this commit is some minor cleanup of existing LESS styles
which were involved or adjacent to styling the new components.

Bug: T405236
Change-Id: I4d7b77a258560a1113d6fbe823d88a2c1021b43c