Check if there are changes before saving statement
We only want to enable the “publish” button if we know that there will
be an edit to save: if we’ve parsed all the values and can tell that
something is different from the saved values – be it the main value, any
qualifiers or references, additions or removals, or rank changes.

For the main snak, this interfaces with the parsedValueStore; qualifiers
and references don’t use that yet, so for now we directly use their data
values as saved in the store. T406887 will change this later.

The practical effect of this is that if you edit a string statement and
only add some whitespace at the end (or beginning), the publish button
will stay disabled until you make some non-whitespace changes.

The tests for qualifiers and references are incomplete; I tried writing
some tests for this and discovered that they failed because changing the
value (or snak type) in the editStatementStore also changed it in the
savedStatementStore – apparently initializeWithStatement() partially
holds references to the same objects still. Given that we don’t yet
support editing qualifiers or references, and the code is definitely
going to change structure once we do, I think it’s okay to not test this
for now and just leave TODOs in the tests.

Bug: T405238
Change-Id: I574f8c1309b1c6bd6bcb369d7202f8e84be50395