Update wdio (v8), wdio-mediawiki (v3), wdio-wikibase (v6.3)
browser.config is gone; get the waitforTimeout and baseUrl from the
browser.options, and move everything else (our custom variables) into
the capabilities. Also pass the user and password into wdio-wikibase’s
initialize() method – compare #72 and [1].

In linkitem.js, change the window size. The viewport width was
previously 1185px under wdio v7, but became 1050 in wdio v8 by default,
which is below MediaWiki core’s @min-width-breakpoint-desktop (1120px)
and thus seems to mess with the display of the tools menu where the link
lives. Increase the size explicitly to make the test pass. (Setting the
*window* width to 1185px actually only results in a viewport width of
1180px, but that’s still enough to fix the test, so I can’t be bothered
to work out what exactly the correct window size would be. There’s also
a setViewport() method, but the docs don’t make it clear whether it can
also be used to increase the size beyond the window or only decrease it
further, and in any case that method is missing a getViewport()
counterpart that would let us clean up the size afterwards.)

[1]: https://github.com/wmde/wikibase-release-pipeline/blob/6c6d285059/test/setup/make-test-settings.ts#L26

Bug: T406844
Change-Id: Ief5eeb23801ab95dcc283ab11a1faaafa020785a