GQL: Determine GraphQL value type by data type
This changes the way the specific value type was determined in the
Value union type. The type was previously resolved by inspecting the
type of value, whereas now it gets the property data type from the
Statement or PropertyValuePair model.

This means that no further changes have to be made to the Value union
type. Additional types of values will be supported as soon as they have
a corresponding entry in the data type definitions.

Bug: T404838
Change-Id: I01f81d3dbcbd70799ad56e912b224549321167c8