Enable the wbui2025 interface for beta feature users
Determine “is this a wbui2025 mobile view?” once, when the
Wbui2025FeatureFlag is called in the lazy-loaded parser options.
In EntityHandler, we read that option (which will then be evaluated
based on $lazyLoad), and pass it through to all the other places that
don’t have access to the parser options directly. In ItemView and
PropertyView, we add a new “placeholder” to the ViewContent, which will
later be copied to the parser output’s extension data; in data
updaters, we can then add the ResourceLoader modules based on that
extension data, without needing another copy of the flag passed in.

Because the "placeholder" variables need to be available in time for
the content view creation, we create a new `getParserOuputOptions`
function in the EntityView interface that the
EntityParserOutputGenerator can call to transfer the options to the
extension data before the content is rendered.

`$wbMobile` is currently a bool|string mix. In a follow-up we can
consider changing this to be an enum.

Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Depends-On: Ifab976f2656b1e76fa2ea2d4c8621ce24c654617
Bug: T407737
Change-Id: I5ebf9743642fafe14b08e2dddbc1d29c90e5576c