Basic implementation of entity-id snak types for wbui2025
Add support for entity-id values and entity-id-typed properties in the
wbui2025 user interface.

Includes support for searching entities and displaying the search
results in a lookup / drop-down.

Support for WikibaseLexeme - Lexemes, Senses and Forms - is excluded
because the extension is not available for all Wikibase installations.
Follow-up changes will add that support

Depends-On: I660dbbfcc43468c1d9b7f339f4432004559b545b
Bug: T403974
Change-Id: I99813bdf9b529892e06356ce84cea5c36a6a6964