Convert wbui2025 to library to make ValueStrategy extensible
wbui2025 uses `ValueStrategies` to handle snak inputs of different
datatypes. Because we also want to support Lexemes, Senses and Forms
without making WikibaseLexeme a dependency of Wikibase, we need the
support for those datatypes to be registered only when the
WikibaseLexeme extension is loaded.

Factory the wbui2025 common code into a library that is its own
`ResourceLoader` module, so that the module can be included from
WikibaseLexeme, and WikibaseLexeme can use the `ValueStrategyFactory`
to register handling of its own datatypes.

Note that this change will turn the `edit` link in wbui2025 from red
to blue for Lexemes, Forms and Senses, even though the actual support
for these datatypes comes only in I0c1de6c7105be8. This is done to
avoid a circular dependency - the patches will hopefully be merged
close to oneanother.

Bug: T409149
Change-Id: I88c9691fceca0422475cf909ad5898a86a6a6964