Fix DateTimeValueCleaner for years in the far future
Since PHP 8.3.25, cal_days_in_month() refuses to operate outside of what
we previously called the “sane range” here, throwing a ValueError
instead of silently giving a slightly wrong result (28 instead of 29
days for February 92000000200). To avoid this error, bring the year into
a PHP-safe range using string operations (so we don’t introduce any
numerical imprecision / overflow issues).

Note that this *changes* the result for affected dates (year 2^31 or
higher, precision day or higher) in all PHP versions, but in a way that
I think is more correct. (And, really, I very much doubt anybody will
notice in practice. As of this writing, there are only 14 distinct time
values in Wikidata in excess of 2^31 years, and none of them are in
February or have a precision higher than year: https://w.wiki/Fwqd.)

Bug: T409269
Change-Id: Ibde22c1e8832b0c46301529eb613fe45dc6f4703