Refactor addReferences and addQualifiers to use editableSnakValue
The Add Reference and Add Qualifier forms have until been using
copy-paste versions of the logic for inputting snak data, meaning that
functionality added to the editableSnakValue (e.g. support for Items)
has been missing.

Refactor the addReferences and addQualifiers views to use the
editableSnakValue component so that we have consistent behaviour
everywhere where we ask the user to select / input a new snak.

Bug: T409147
Change-Id: I9bad603aab37b9aa2c1d99365ee3d3376a6a6964