Add strict types to LanguageFallbackChainFactory(Test)
In accordance with ADR 11 [1]; my specific motivation here is avoiding
any implicit casts between LanguageFallbackMode and int. (PHP doesn’t
currently perform any such conversions, but the RFC [2] doesn’t rule
them out for the future.)

In getBabel(), we need to cast the level to string (PHP may have cast
the array key to int if it was a number). This is not covered by
testNewFromUserAndLanguageCode() (which only tests an anonymous user),
but was revealed by SpecialPageFatalTest in CI.

[1]: https://doc.wikimedia.org/Wikibase/master/php/adr_0011.html
[2]: https://wiki.php.net/rfc/enumerations#auto-scalar_conversion

Change-Id: I5be0f6f88bb77d738ef56a13b21d692f37c62637