Introduce modal for adding value to statements
Currently, clicking the 'add value' button in the statement modal
creates a new empty value directly in the statement modal for the user
to fill out. After a UX review, it was determined that this add value
flow is not consistent with the rest of the user experience, in which
all input requests are served through fresh modal dialogs.

Change the 'add value' functionality so that a new 'Add value' modal
is displayed for inputting data about the new value.

Bug: T406878
Change-Id: Id02e08c95fd04bf26ce2ae4cf7b8fdb1ee7c3cbc