Use valueStrategy pattern to manage componens in editableSnakValue
The EditableSnakValue currently includes an if/else-if/else block to
determine which codex component to show for the snak, and includes all
the logic for all snak types.

Factor out the different snak datatypes (string, lookup) into their
own components to improve readability and encapsulation. Use Vue's
`<component :is=...>` and the ValueStrategy pattern to make the
component loading more flexible / extensible.

Bug: T409148
Change-Id: I5b24af4486702189df73e90e5a49954e6a6a6964