Disable unsupport datatypes
In I88c9691fceca0422 we changed the list of supported datatypes
to a list of unsupported datatypes, to make it easier to include
datatypes that are part of extensions (e.g. WikibaseLexeme)
without adding additional hooks to register the supported datatypes.

The original change missed two datatypes which are present in Wikidata
but not yet supported in MEX / wbui2025 - musical notation and
mathematical expressions.

Add the as-yet-unsupported datatypes to the unsupported datatypes
list.

Change-Id: Ic8acb7e4a64c9d54b81d8aa76c79e6936a6a6964