Update wdio-mediawiki (v5), wdio-wikibase (v7)
Pull in the new versions and rewrite all the WebdriverIO browser test
files to ECMAScript module syntax. (We can drop 'use strict'; as strict
mode is the default in ESM.) Also pull in an update to the termbox
submodule to the same effect.

The data-bridge browser test .eslintrc.js apparently can’t be changed to
module syntax (not supported by legacy ESLint – note that “modern”
ESLint supports an entirely different config file), so rename it to .cjs
so it still works with the new "type": "module" package.json.

Also add a missing `await` on a waitForModuleState call in
dataBridge.page.js (oops!).

Bug: T406844
Change-Id: I317e41220bca0d27297ed9b20b7ada77a6e46c71
Depends-On: Iffa95bfaaa6ee5723fbdd52a5a6ab586c19c8efd
Depends-On: I79d740a418c4dbf83f2baa8012fcb982d078522b