Use EntityDiff for EditSummaryHelper
Prepare the edit summary based on the diff of the two entities, rather
than the ChangeOpResult; this has more information available for us.

The test cases in EditSummaryHelperTest become somewhat simpler because
two complications fall away: the arbitrary nesting of ChangeOpResults
(the EntityDiff is always structured), and the concept of results that
didn’t change the entity (no diff in that case).

Bug: T411247
Change-Id: Ia0d71e037e041a57dfa8a4e57826511b08e8e4f1