Move more summary generation into EditSummaryHelper
I think it makes sense for this code to live in this class; more
importantly, for good statement edit summaries we’ll need to be able to
control the “module” of the summary, which can’t be changed later. (We
could add a setter for it to Summary, of course, but I think it’s nicer
for EditSummaryHelper to create a Summary with the right module
directly.)

Since this effectively adds a copy of ModifyEntity::createSummary() (to
add the user summary from the params to the Summary), introduce a
constant on ModifyEntity for the name of this parameter, so that it’s
possible to find this reference by looking for usages of the constant.
(There are also other API modules with this parameter, and references to
that parameter of those modules elsewhere, but those are all defined
independently – SetClaim etc. don’t extend ModifyEntity – so those stay
unmodified and don’t use the new constant.)

Bug: T411247
Change-Id: I377d95379cca22c11266e325694e1f6c66726d5f