Refactor EditSummaryHelper for different modules
For upcoming improvements to statement change edit summaries,
EditSummaryHelper will need to be able to return edit summaries with
different modules, which doesn’t fit well with its current structure of
creating a summary once and then modifying (“preparing”) it. Refactor it
so each method can return a fresh summary.

This partially belongs to change I377d95379c, and partially to change
I99122c508f, but I’m committing it separately so it’s easier to review.

Also fix some whitespace in the test case names.

Bug: T411247
Change-Id: I17a35c026b3f02a921f33a6f10ef07fdcdd233bf