Use ClaimSummaryBuilder in EditSummaryHelper
If the diff only affects the statements, and the claims (statements)
diff only contains one statement change, analyze that using the
ClaimSummaryBuilder which is also used by SetClaim and use that edit
summary. Otherwise, continue to use a generic edit summary. Later, we’ll
want to extend this for at least edits to multiple statements for the
same property, which we expect to see for MEX.

One issue with ClaimSummaryBuilder is that if you edit a qualifier, it
reports it as “Changed 2 qualifiers of claim”, because it sees a
qualifier diff with two diff ops (remove old qualifier, add new
qualifier) rather than one (change qualifier value). This issue is
visible in some old Wikidata edit summaries (see T410130#11383715 for
details), but not since April 2014, because the “update-qualifiers” and
“update-references” summaries have been effectively unused since then:
SetClaim passes the summary from the ClaimSummaryBuilder into the
ChangeOp, which in turn will reset the action to a plain “update” (while
keeping the number 2 in the autocomment args invisibly in place – you
can see it in e.g. API response [1] for edit [2]).

[1]: https://www.wikidata.org/w/api.php?action=query&revids=2429456822&prop=revisions&formatversion=2
[2]: https://www.wikidata.org/w/index.php?diff=2429456822

Bug: T411247
Change-Id: I99122c508ff2ab01aaa28539ef84d5cba755ff5f