Further improve wbeditentity summaries for terms
For the case where exactly one language is changed (which I assume is
common even if the mobile termbox allows changing multiple languages at
once), augment EditSummaryHelper to reuse the more specific summaries of
the API modules wbsetlabel, wbsetdescription, wbsetaliases and
wbsetlabeldescriptionaliases. The code to analyze the diff and reproduce
the edit summaries gets a bit finicky – it doesn’t help that the
existing code sets the edit summaries in multiple places: some happen in
ChangeOp::apply(), but there’s also SetAliases::adjustSummary() and
SpecialSetLabelDescriptionAliases::getSummaryForLabelDescriptionAliases()
– but overall I think it’s not *too* bad and should be worth it. Also,
we can test this code pretty thoroughly in EditSummaryHelperTest.

This is especially nice for removing terms in unsupported languages
(T403097), because (assuming change Ie034ae9ab8 is merged) you can
*only* use wbeditentity for those edits, so there’s no way to get ”nice”
edit summaries without this change – but with this change, you get them
“for free”.

In provideItemIdParamsAndExpectedSummaryPatternForEditEntity() and
testEditEntity_producesCorrectSummary(), also fix the preg_quote()
usage. If no delimiter is specified, preg_quote() does not escape it;
also, assertRevisionSummary() doesn’t add a delimiter: it expects the
pattern / regex to already be a full preg_match() input (if it’s a
string). This means that the tests were actually using the surrounding
//s of the edit summary as the regex delimiter, rather than as part of
the summary to expect, and trying to assert an edit summary with an auto
summary after the auto comment didn’t work properly. Fix this by moving
the preg_quote() into testEditEntity_producesCorrectSummary() and making
the data provider just provide fixed strings.

Bug: T411247
Bug: T403097
Bug: T370597
Change-Id: I4747f9da7d837c9d6f4ad16a7940b3e687f05fe6