Avoid covering modal save / cancel buttons with messages
The statusMessage overlay currently displays over the top of whatever
is at the bottom of the screen in MEX / wbui2025. When a statement has
recently been submitted, this causes the status messages to cover the
publish and cancel buttons of the edit statement form (if it is opened
again). This is possibly confusing for users and certainly an issue in
the end-to-end tests - the tests are currently delayed waiting for the
status messages to disappear.

Clear any messages when the user context changes (i.e. when the modal
form opens up).

Bug: T409828
Change-Id: I67fe49baf8726816cb5fd8edcf753a0c6a6a6964