Fix padding on dropdown menu items
Menu items inside Codex Select lists currently appear with a margin
down their left-hand side that is not part of the design. This
appears to come from the Minerva skin [1] and addresses any `ul`
element in a `.content` section.

Override the style in `editStatementGroup.vue` and `modalOverlay.vue`
(the top-level wbui2025 components) to set the left-padding on `ul`
elements in MEX / wbui2025 to 0. While we are there, also remove the
margin underneath the list items that is set by the same Minerva
stylesheet.

1. https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/skins/MinervaNeue/+/refs/heads/master/resources/skins.minerva.styles/content/links.less

Bug: T410408
Change-Id: I02d0d65263ddde729226be0f997b7ddf6a6a6964