Fix cutoff publish / cancel buttons on mobile
The modal overlay is currently sized according to the codex token
`size-viewport-height-full`, which works fine on desktop. Unfortunately
it is a long-standing and apparently well-known[1] issue that the
`100vh` value is bigger than the viewable screen on mobile devices -
it does not take into accout UI chrome that mobile browsers may have
present on the screen in various device states.

Viewport variants [2] were introduced in the 2021 CSS working draft
and have been supported in major browsers since 2022. We expect the
units to be supported by the majority of browsers that are able to
access the wbui2025 dynamic features [3].

Nevertheless, since in this case `100%` also makes the modal take
up the full screen and is universally supported, we use that here
instead.

1. https://web.dev/blog/viewport-units
2. https://www.w3.org/TR/css-values-4/#viewport-variants
3. https://caniuse.com/viewport-unit-variants

Bug: T411001
Change-Id: Ib2ec59ee2d7941a3f2ec17b3ac5560c36a6a6964