Indicate errors on input fields
Codex will only display the error if the field is out of
focus, so we don't need to remove the class once the input
is focused.

This also ensures that, when changing a value snak to a
novalue / unknown value one and back, the last "complete"
value will be restored. When entering a new value, this will
discard the input.

Per [0], for this to work, I needed to convert useEditSnakStore
to a Setup Store[1]. As the store itself is not available
via this in getters, we need to store a reference to it in
the wrapper function.

[0]: https://github.com/vuejs/pinia/discussions/794
[1]: https://pinia.vuejs.org/core-concepts/#Setup-Stores

Bug: T408928
Change-Id: I7d37871c1d3240999dbcec615ac817927cc2477f