Make editStringDatatypes Cypress tests run offline
The Cypress tests for editStringDatatypes currently stub some of the
calls to Commons, but not all of them, and they rely on the data
being available in Commons for Wikibase to be able to validate the
values.

Since we want our tests to run even if the items in commons
change or if commons is offline when we want to run tests, update
the names of the test items so that they do not match those in
commons.

In order to test the Commons-backed string datatype functionality
of the editStatement form without relying on items of the matching
datatypes (which Wikibase would then try and validate), we create
items with plain string properties, and change the statement data
at runtime so that the Vue UI thinks that it is working with
tabular-data or geo-shape properties.

Bug: T410998
Change-Id: Idc11d11edd19ae5197dff6a0f1d1168b6a6a6964