Fix broken entity search when switching from novalue to custom value
If an entity-valued statement is saved as 'novalue', editing the
statement and switching the snak type to 'custom value' results in a
broken entity search - no search is triggered and a value cannot be
saved.

Set the initial value for the lookupInput to the empty string, even
if the snak has previously been `novalue` or `unknownvalue` (where the
lookupInput value would otherwise be initialised as `null`).

Bug: T411236
Change-Id: Id5e2abe422f37979fe9927883d522d516a6a6964