Update top-level statement sections as statements change
The current implementation of the top-level statementSections
component renders the component using php-vuejs-templating, but does
not hydrate a top-level statementSections component. This means that
the statement sections in the item view are static and do not get
updated when new properties are added or all the statements for an
existing property are removed.

Mount the top-level statementSections components from entityViewInit
and make them respond to the appearance of new properties and the
removal of old properties.

Note that in this implementation, when a user clicks the 'add
statement' button of a particular statements section, the front-end
has no mechanism for verifying that the new property actually belongs
in the section to which it has been added. Statements may be
incorrectly grouped until the page is refreshed and loaded again from
the server. This matches the current behaviour of the deskop interface.
The task for addressing this is T411369, but it is not yet clear that
this will be adopted into the product scope.

Bug: T411237
Change-Id: Ie7ee7ad19469df4dc39406fddbf11bf06a6a6964