Add edit summaries for editing multiple statements
Add two summaries for edits that can only be made via wbeditentity:
editing multiple statements at once. (wbremoveclaims can remove multiple
statements at once, but wbsetclaim only takes a single statement to
edit.) One summary is for the case where all the statements have the
same property ID (the MEX / wbui2025 case): here we can show that
property ID in the edit summary (but not the statement values). The
other summary is for the remaining case: statements with different
properties have been edited, and we just show the number of properties
affected.

Bug: T411247
Change-Id: I68f586ea1ac7b8b8e239e21f9102c5d259ae0ad9