Detect additional statement changes in EditSummaryHelper
In addition to changes of a statement, also support additions and
removals of single statements. The multiple-statement case remains TBD
(see change I68f586ea1a).

In both cases, the summary args are double-nested, an outer array
containing one inner associative array (I tested this for the actual API
modules with a breakpoint in EntitySavingHelper::attemptSaveEntity()).
For the case of adding a statement, the resulting Summary object is
actually not *quite* the same as that created by wbsetclaim – there you
end up with summaryArgs like this:

    [ 0 => [ 'P1' => /* snak */ ], 1 => '' ]

because the first autosummary arg is added by ClaimSummaryBuilder, and
then the summary is passed into ChangeOpStatement, which calls
ChangeOpBase::updateSummary() without any args, and ChangeOpBase then
adds the default $args ('') to the autosummary args. But this empty
string gets stripped out again by SummaryFormatter::formatAutoSummary(),
so once the edit summary reaches MediaWiki core and the database,
there’s no trace of the extra arg anymore; IMHO this is a harmless bug
in the existing code and not worth replicating in EditSummaryHelper.

If we see an unexpected type of DiffOp that we can’t make sense of, log
a warning for further investigation and fall back to the generic edit
summary. I’m a bit worried about the potential of this warning spamming
logstash, so include a comment to the effect that anyone who comes
across it is free to remove the code logging the warning, as long as a
Phabricator task has been filed for our attention. (If we don’t see the
warning in production, we can probably turn this into an assertion
sooner or later.)

Bug: T411247
Change-Id: I077b6a774e8ac08d58de5c3e99c36f79d4072c1e