Add indicator icons to snaks
Create structure in the pinia store for wbui2025 / MEX to store HTML
corresponding to "indicators" - especially warning / notification
icons - that can be associated with snaks and displayed in the snak
view.

We expect that the WikibaseQualityConstraints and Score extensions
will use the indicator mechanism. In a follow-up patch, we will make
the indicators clickable - clicking will display a pop-up containing
more information about the indicator.

Bug: T411604
Change-Id: I6ee88a8de409502517c024538abc199b6a6a6964