Add popover for indicators
Extend the indicator support in the previous patch (I6ee88a8de4095) to
make the indicators clickable. On click, a popover is displayed. The
content of the popover must be stored in the pinia store, and currently
consists of a title, a Codex icon name, and some HTML for the content.

Bug: T411608
Change-Id: I3965a5a9f60ba80f2c7467a5b80e49406a6a6964