Search: Increase random string length in test
The test relies on uniquely identifying an item or a property by the
prefix of its label which includes a suffix of 10 randomly generated
characters. Since the test removes 8 characters at the end, this
results in a relatively high chance (1 in 26^2?) of accidentally
matching another item or property's label. Increasing the length of the
search term used in the prefix search significantly lowers this risk.

Bug: T411264
Change-Id: I72bbcb517bf8b4491526cdedab08829a11de14f9