Support time values in wbui2025
Enable support for the 'time' datatype in wbui2025. Time values can be
added to statements, references and qualifiers, and can be edited.

The implementation introduces a popup when the snak value field is
selected. Precision and Calendar can be manually adjusted or left at
their automatically-determined defaults.

The current CdxPopover implementation uses the `@floating-ui/vue`
library to set the width of the floating component. The implementation
uses `autoUpdate` which, per the docs [1], does not update the width
in all situations. In particular, when the wbui2025 code fetches the
calculated precision and calendar values from the server, the width
of those elements in the popover change which can result in a
horizontal scrollbar inside the popover content area. T414917 should
address this issue.

1. https://floating-ui.com/docs/vue#manual-updating

Bug: T409453
Change-Id: I0e865b6079d21b8a1a2574b311ed11d56a6a6964