Indicate which properties are not yet supported in addstatement
We are progressively implementing datatypes in the wbui2025 mobile
interface. This means that properties of some datatypes should not
be able to be added to statements in the mobile interface since
users will not then be able to edit them.

Mark the items in the property lookup results as disabled if the
datatype is not yet supported, and annotate them with the text
'not supported on mobile'.

Bug: T412719
Change-Id: I232775c3bddde91440fedb3f919c01a66a6a6964